gh-127085: use atomic updating for `exports` and add protection for reading.
